Redesign Setup View controller. Add global close button to the navigation bar.
